refactor/#20 : JWT 관련 코드 폴더 변경
